Made one more small change to get closer to compilation.
Then struggled to progress because I couldn't get tinyxml2 to link properly.

+++ src/tools/hidpp10-load-temp-profile.cpp
@@ -116,7 +116,10 @@ int main (int argc, char *argv[])
 	profxml.read (doc.RootElement (), profile, entry, macros);
 
 	try {
-		HIDPP::Address macro_address { 0, 0, (profile_format->size ()+1)/2 };
+		// HIDPP::Address macro_address { 0, 0, (profile_format->size ()+1)/2 };
+		HIDPP::Address macro_address { 0, 0, static_cast<unsigned int>((profile_format->size ()+1)/2) };
+		// ^ Had to insert static_cast for this to compile under macOS
+
 		for (unsigned int i = 0; i < profile.buttons.size (); ++i) {
 			auto &button = profile.buttons[i];
 			if (button.type () == HIDPP::Profile::Button::Type::Macro) {
